pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "hooke-surveyor"
version = "0.1.0"
description = "Territorial surveyor for Hooke Park — crown segmentation, tree snapshots, continuous monitoring at H3 res-13"
requires-python = ">=3.11"
dependencies = [
    "geopandas>=0.14",
    "h3>=4.0",
    "numpy>=1.26",
    "pandas>=2.1",
    "rasterio>=1.3",
    "shapely>=2.0",
    "scipy>=1.12",
    "pyproj>=3.6",
    "google-cloud-storage>=2.14",
    "fiona>=1.9",
]

[project.optional-dependencies]
cuda = ["cupy-cuda12x>=13.0"]
dev = ["pytest>=8.0", "ruff>=0.3"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.pytest.ini_options]
testpaths = ["tests"]
